add outofproc proj to sln
