+++ Worker.Extension.Sql/src/SqlOutputAttribute.cs
@@ -3,19 +3,18 @@
 
 using Microsoft.Azure.Functions.Worker.Extensions.Abstractions;
 
-namespace Microsoft.Azure.Functions.Worker.Sql
+namespace Microsoft.Azure.Functions.Worker.Extension.Sql
 {
-    public sealed class SqlInputAttribute : InputBindingAttribute
+    public class SqlOutputAttribute : OutputBindingAttribute
     {
         /// <summary>
         /// Constructs a new instance.
         /// </summary>
         /// <param name="commandText">The text of the command.</param>
         /// <param name="connectionStringSetting">The name of the app setting where the SQL connection string is stored.</param>
-        public SqlInputAttribute(string commandText, CommandType commandType, string connectionStringSetting)
+        public SqlOutputAttribute(string commandText, string connectionStringSetting)
         {
             this.CommandText = commandText;
-            this.CommandType = commandType;
             this.ConnectionStringSetting = connectionStringSetting;
         }
 
@@ -31,7 +30,7 @@ namespace Microsoft.Azure.Functions.Worker.Sql
         public string ConnectionStringSetting { get; set; }
 
         /// <summary>
-        /// For an input binding, either a SQL query or stored procedure that will be run in the database referred to in the ConnectionString.
+        /// For an output binding, the table name.
         /// </summary>
         public string CommandText { get; }
 
@@ -39,17 +38,6 @@ namespace Microsoft.Azure.Functions.Worker.Sql
         /// Specifies whether <see cref="CommandText"/> refers to a stored procedure or SQL query string.
         /// Use <see cref="CommandType.StoredProcedure"/> for the former, <see cref="CommandType.Text"/> for the latter
         /// </summary>
-        public CommandType CommandType { get; set; } = CommandType.Text;
-
-        /// <summary>
-        /// Specifies the parameters that will be used to execute the SQL query or stored procedure specified in <see cref="CommandText"/>.
-        /// Must follow the format "@param1=param1,@param2=param2". For example, if your SQL query looks like
-        /// "select * from Products where cost = @Cost and name = @Name", then Parameters must have the form "@Cost=100,@Name={Name}"
-        /// If the value of a parameter should be null, use "null", as in @param1=null,@param2=param2".
-        /// If the value of a parameter should be an empty string, do not add anything after the equals sign and before the comma,
-        /// as in "@param1=,@param2=param2"
-        /// Note that neither the parameter name nor the parameter value can have ',' or '='
-        /// </summary>
-        public string Parameters { get; set; }
+        public System.Data.CommandType CommandType { get; set; } = System.Data.CommandType.Text;
     }
 }
