--- Worker.Extensions.Sql/src/Worker.Extensions.Sql.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<AssemblyName>Microsoft.Azure.Functions.Worker.Extensions.Sql</AssemblyName>
-		<RootNamespace>Microsoft.Azure.Functions.Worker.Extensions.Sql</RootNamespace>
-		<Description>Sql extensions for .NET isolated functions</Description>
-
-		<!--Version information-->
-		<VersionPrefix>4.0.0</VersionPrefix>
-		<VersionSuffix>-preview2</VersionSuffix>
-
-		<!--Temporarily opting out of documentation. Pending documentation-->
-		<GenerateDocumentationFile>false</GenerateDocumentationFile>
-	</PropertyGroup>
-
-	<Import Project="..\..\builds\Extensions.props" />
-
-	<ItemGroup>
-		<PackageReference Include="Microsoft.Azure.Functions.Worker" />
-	</ItemGroup>
-
-</Project>
